requirements.txt
# Core Web Framework
streamlit==1.31.0
flask==3.0.0
flask-cors==4.0.0

# Google Cloud APIs
google-generativeai==0.3.2
google-cloud-vision==3.7.0

# NLP Libraries
nltk==3.8.1
spacy==3.7.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Machine Learning & Embeddings
rank-bm25==0.2.2
sentence-transformers==2.3.1
langchain==0.1.6
langchain-community==0.0.19
faiss-cpu==1.7.4

# Data Processing & Visualization
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
openpyxl==3.1.2

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0

